backend/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
greenlet==3.0.3
idna==3.10
iniconfig==2.0.0
packaging==24.1
playwright==1.47.0
pluggy==1.5.0
psutil==6.0.0
psycopg==3.2.3
psycopg2-binary
py==1.11.0
pyee==12.0.0
PyJWT
pytest==8.3.3
pytest-base-url==2.1.0
pytest-playwright==0.5.2
pytest-xprocess==1.0.2
python-slugify==8.0.4
requests==2.32.3
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.3
Flask == 3.0.3